+++ src/test/java/roomescape/client/TossPaymentClientTest.java
@@ -1,115 +1,97 @@
 package roomescape.client;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
+import org.apache.hc.client5.http.ConnectTimeoutException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.client.RestClientTest;
-import org.springframework.context.annotation.Import;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.test.web.client.MockRestServiceServer;
-import roomescape.client.dto.PaymentConfirmResultDto;
-import roomescape.config.PaymentClientConfig;
-import roomescape.domain.payment.Payment;
-import roomescape.domain.reservation.Reservation;
-import roomescape.exception.PaymentConfirmClientException;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.HttpHeaders;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+import org.springframework.web.client.ResourceAccessException;
+import org.springframework.web.client.RestClient;
 import roomescape.exception.PaymentConfirmServerException;
-import roomescape.service.dto.PaymentConfirmDto;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import java.net.SocketTimeoutException;
+import java.util.Base64;
+
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.springframework.test.web.client.match.MockRestRequestMatchers.*;
-import static org.springframework.test.web.client.response.MockRestResponseCreators.withStatus;
-import static org.springframework.test.web.client.response.MockRestResponseCreators.withSuccess;
 
-@RestClientTest(TossPaymentClient.class)
-@Import(PaymentClientConfig.class)
-class TossPaymentClientTest {
+@ExtendWith(SpringExtension.class)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
+@ActiveProfiles("test")
+public class TossPaymentClientTest {
 
-    @Autowired
-    private ObjectMapper objectMapper;
+    public static final String CONNECT_TIMEOUT_TEST_URL = "http://10.255.255.1";
+    public static final String READ_TIMEOUT_TEST_URL = "https://httpstat.us/200?sleep=100000";
 
     @Autowired
-    private MockRestServiceServer mockServer;
+    @Qualifier("tossRestClient")
+    private RestClient tossRestClient;
 
-    @Autowired
-    private TossPaymentClient tossPaymentClient;
-
+    @DisplayName("결제 예외 핸들링 테스트 - INVALID_API_KEY가 발생할 경우")
     @Test
-    @DisplayName("UNAUTHORIZED_KEY 예외를 받으면 PaymentConfirmServerException를 던진다")
-    void test1() throws JsonProcessingException {
-        // given
-        PaymentConfirmDto requestDto = new PaymentConfirmDto("paymentKey", "orderId", 10000L);
-        String errorResponseBody = """
-                    {
-                      "code": "UNAUTHORIZED_KEY",
-                      "message": "인증되지 않은 시크릿 키 혹은 클라이언트 키 입니다."
-                    }
-                """;
-
-        mockServer.expect(requestTo("https://api.tosspayments.com/v1/payments/confirm"))
-                .andExpect(method(HttpMethod.POST))
-                .andExpect(content().json(objectMapper.writeValueAsString(requestDto)))
-                .andRespond(
-                        withStatus(HttpStatus.UNAUTHORIZED)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .body(errorResponseBody)
-                );
-
-        // when & then
-        assertThatThrownBy(() -> tossPaymentClient.confirmPayment(requestDto))
-                .isInstanceOf(PaymentConfirmServerException.class);
+    void paymentExceptionTest() {
+        assertThatThrownBy(
+                () -> tossRestClient.post()
+                        .uri("/v1/payments/confirm")
+                        .header("Authorization", "Basic " +
+                                Base64.getEncoder().encodeToString("test_gsk_docs_OaPz8L5KdmQXkzRz3y47BMw6:".getBytes()))
+                        .header("TossPayments-Test-Code", "INVALID_API_KEY")
+                        .retrieve()
+                        .toBodilessEntity()
+        ).isInstanceOf(PaymentConfirmServerException.class);
     }
 
+    @DisplayName("결제 예외 핸들링 테스트 - 유효하지 않은 Secret Key인 걍우")
     @Test
-    @DisplayName("사용자에게 보여줘도 되는 메시지의 예외는 그대로 PaymentConfirmClientException를 던진다.")
-    void test2() throws JsonProcessingException {
-        // given
-        PaymentConfirmDto requestDto = new PaymentConfirmDto("paymentKey", "orderId", 10000L);
-
-        String errorResponseBody = """
-                    {
-                      "code": "ALREADY_PROCESSED_PAYMENT",
-                      "message": "이미 처리된 결제 입니다."
-                    }
-                """;
-
-        mockServer.expect(requestTo("https://api.tosspayments.com/v1/payments/confirm"))
-                .andExpect(method(HttpMethod.POST))
-                .andExpect(content().json(objectMapper.writeValueAsString(requestDto)))
-                .andRespond(
-                        withStatus(HttpStatus.BAD_REQUEST)
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .body(errorResponseBody)
-                );
-
-        // when & then
-        assertThatThrownBy(() -> tossPaymentClient.confirmPayment(requestDto))
-                .isInstanceOf(PaymentConfirmClientException.class);
+    void paymentExceptionTest2() {
+        assertThatThrownBy(
+                () -> tossRestClient.post()
+                        .uri("/v1/payments/confirm")
+                        .header("Authorization", "Basic " +
+                                Base64.getEncoder().encodeToString("testFail_gsk_docs_OaPz8L5KdmQXkzRz3y47BMw6:".getBytes()))
+                        .retrieve()
+                        .toBodilessEntity()
+        ).isInstanceOf(PaymentConfirmServerException.class);
     }
 
+    @DisplayName("결제 예외 핸들링 테스트 - 유효하지 않은 Secret Key인 걍우")
     @Test
-    @DisplayName("정상 흐름 진행 테스트")
-    void test3() throws JsonProcessingException {
-        // given
-        PaymentConfirmDto requestDto = new PaymentConfirmDto("paymentKey", "orderId", 10000L);
+    void paymentExceptionTest3() {
+        assertThatThrownBy(
+                () -> tossRestClient.post()
+                        .uri("/v1/payments/confirm")
+                        .header(HttpHeaders.AUTHORIZATION, "Basic " +
+                                Base64.getEncoder().encodeToString("test_gsk_docs_OaPz8L5KdmQXkzRz3y47BMw6".getBytes()))
+                        .retrieve()
+                        .toBodilessEntity()
+        ).isInstanceOf(PaymentConfirmServerException.class);
+    }
 
-        Payment payment = new Payment(requestDto.paymentKey(), requestDto.orderId(), requestDto.amount(), new Reservation());
+    @DisplayName("Connect Timeout 테스트")
+    @Test
+    void connectTimeoutExceptionTest() {
+        assertThatThrownBy(
+                () -> tossRestClient.post()
+                        .uri(CONNECT_TIMEOUT_TEST_URL)
+                        .retrieve()
+                        .toBodilessEntity()
+        ).isInstanceOf(ResourceAccessException.class)
+                .hasCause(new ConnectTimeoutException("Connect to http://10.255.255.1:80 failed: Connect timed out"));
+    }
 
-        mockServer.expect(requestTo("https://api.tosspayments.com/v1/payments/confirm"))
-                .andExpect(method(HttpMethod.POST))
-                .andExpect(content().json(objectMapper.writeValueAsString(requestDto)))
-                .andRespond(withSuccess(objectMapper.writeValueAsString(payment), MediaType.APPLICATION_JSON));
-
-        // when
-        PaymentConfirmResultDto result = tossPaymentClient.confirmPayment(requestDto);
-
-        //then
-        assertThat(result.paymentKey()).isEqualTo(payment.getPaymentKey());
-        assertThat(result.orderId()).isEqualTo(payment.getOrderId());
-        assertThat(result.totalAmount()).isEqualTo(payment.getTotalAmount());
+    @DisplayName("Read Timeout 테스트")
+    @Test
+    void readTimeoutExceptionTest() {
+        assertThatThrownBy(
+                () -> tossRestClient.post()
+                        .uri(READ_TIMEOUT_TEST_URL)
+                        .retrieve()
+                        .toBodilessEntity()
+        ).isInstanceOf(ResourceAccessException.class)
+                .hasCause(new SocketTimeoutException("Read timed out"));
     }
 }
