+++ src/test/resources/application-test.properties
@@ -14,6 +14,6 @@ spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.defer-datasource-initialization=false
 
-toss.payment.confirm.connect-timeout=3000
-toss.payment.confirm.read-timeout=30000
+toss.payment.confirm.connect-timeout=1000
+toss.payment.confirm.read-timeout=1000
 toss.payment.confirm.secretKey=test_gsk_docs_OaPz8L5KdmQXkzRz3y47BMw6:
